[AC-5901] Hoist function up one more level.
