+++ web/impact/impact/v1/views/analyze_judging_round_view.py
@@ -1,11 +1,14 @@
 # MIT License
 # Copyright (c) 2017 MassChallenge, Inc.
 
+from django.db.models import Count
 from itertools import chain
 from accelerator.models import (
     Application,
     CriterionOptionSpec,
+    JudgePanelAssignment,
     JudgingRound,
+
 )
 from rest_framework.response import Response
 from impact.v1.views.impact_view import ImpactView
@@ -52,6 +55,21 @@ class AnalyzeJudgingRoundView(ImpactView):
         self.apps = Application.objects.filter(
             application_status="submitted",
             application_type=self.instance.application_type)
-        analyses = [OptionAnalysis(option, self.apps, self.instance).analyses()
+        application_counts = self.judge_to_count()
+        analyses = [OptionAnalysis(option,
+                                   self.apps,
+                                   self.instance,
+                                   application_counts).analyses()
                     for option in options]
         return Response({"results": list(chain.from_iterable(analyses))})
+
+    def judge_to_count(self):
+        assignments = JudgePanelAssignment.objects.filter(
+            scenario__judging_round=self.instance)
+        judge_assignment_counts = assignments.annotate(
+            assignment_count=Count("panel__applications")).values_list(
+                "judge_id", "assignment_count")
+        results = {}
+        for judge_id, assignment_count in judge_assignment_counts:
+            results[judge_id] = results.get(judge_id, 0) + assignment_count
+        return results
